refactor: 상세페이지 filter부분을 컴포넌트화
